Move segmentTrianglesIntersection to _glutils
