+++ silx/gui/plot3d/items/mesh.py
@@ -35,6 +35,7 @@ __date__ = "17/07/2018"
 import logging
 import numpy
 
+from ... import _glutils as glu
 from ..scene import primitives, utils, function
 from ..scene.transform import Rotate
 from .core import DataItem3D, ItemChangedType
@@ -168,7 +169,7 @@ class _MeshBase(DataItem3D):
                 _logger.warning("Unsupported draw mode: %s" % mode)
                 return None
 
-        trianglesIndices, t, barycentric = utils.segmentTrianglesIntersection(
+        trianglesIndices, t, barycentric = glu.segmentTrianglesIntersection(
             rayObject, triangles)
 
         if len(trianglesIndices) == 0:
@@ -494,7 +495,7 @@ class _CylindricalVolume(DataItem3D):
         positions = self._mesh.getAttribute('position', copy=False)
         triangles = positions.reshape(-1, 3, 3)  # 'triangle' draw mode
 
-        trianglesIndices, t = utils.segmentTrianglesIntersection(
+        trianglesIndices, t = glu.segmentTrianglesIntersection(
             rayObject, triangles)[:2]
 
         if len(trianglesIndices) == 0:
